Use TestBase from folio-service-tools-test
